requirements.txt
browser-use==0.5.11
playwright==1.54.0
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
apscheduler==3.11.0
aiosqlite==0.21.0
pydantic==2.11.7
SQLAlchemy==2.0.43